Test if compilation is possible

Test if compilation appears possible. This is used in some examples, and tries compiling a trivial C program with pkgbuild. Results are cached between runs within a session so this should be fast to rely on.

can_compile(verbose = FALSE, refresh = FALSE)

Arguments

verbose

Be verbose when running commands?

refresh

Try again to compile, skipping the cached value?

Value

A logical scalar

Details

We use pkgbuild in order to build packages, and it includes a set of heuristics to locate and organise your C compiler. The most likely people affected here are Windows users; if you get this ensure that you have rtools installed. Using pkgbuild::find_rtools() with debug = TRUE may be helpful for diagnosing compiler issues.

Examples

can_compile() # will take ~0.1s the first time
#> Loading required namespace: pkgbuild
#> [1] TRUE
can_compile() # should be basically instantaneous
#> [1] TRUE